change: move some basic objects out and madk a new h&cpp
